requirements.txt
aiogram==3.7.0
google-generativeai==0.7.2
duckduckgo-search==6.2.6
aiosqlite==0.20.0
python-dotenv==1.0.1
